+++ src/main/kotlin/dev/slint/ideaplugin/ide/actions/PreviewAction.kt
@@ -6,7 +6,6 @@ import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.CommonDataKeys
 import com.intellij.openapi.actionSystem.CommonDataKeys.VIRTUAL_FILE
 import com.intellij.platform.lsp.api.LspServer
-import dev.slint.ideaplugin.ide.lsp.SlintLspServer
 import dev.slint.ideaplugin.ide.lsp.requests.PreviewMessageRequest
 import dev.slint.ideaplugin.lang.SlintLanguage
 import kotlin.io.path.Path
@@ -19,12 +18,12 @@ internal class PreviewAction(private val notification: Notification? = null) :
         e.presentation.isEnabledAndVisible = psiFile.language.isKindOf(SlintLanguage.INSTANCE)
     }
 
-    override fun actionPerformed(e: AnActionEvent, servers: List<SlintLspServer>) {
+    override fun actionPerformed(e: AnActionEvent, servers: List<LspServer>) {
         val virtualFile = e.getData(VIRTUAL_FILE) ?: return
         val uriFile = Path(virtualFile.path).toUri()
 
         val request = PreviewMessageRequest(servers.first(), uriFile.toString(), "")
-        (servers.first() as LspServer).requestExecutor.sendRequestSync(request)
+        servers.first().requestExecutor.sendRequestSync(request)
 
         notification?.expire()
     }