+++ src/main/kotlin/dev/slint/ideaplugin/ide/actions/PreviewComponentAction.kt
@@ -5,7 +5,6 @@ import com.intellij.notification.Notification
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.CommonDataKeys
 import com.intellij.platform.lsp.api.LspServer
-import dev.slint.ideaplugin.ide.lsp.SlintLspServer
 import dev.slint.ideaplugin.ide.lsp.requests.PreviewMessageRequest
 import kotlin.io.path.Path
 
@@ -15,12 +14,12 @@ internal class PreviewComponentAction(
 ) :
         LspAction("Show Component Preview", null, AllIcons.Actions.ShowCode)
 {
-    override fun actionPerformed(e: AnActionEvent, servers: List<SlintLspServer>) {
+    override fun actionPerformed(e: AnActionEvent, servers: List<LspServer>) {
         val virtualFile = e.getData(CommonDataKeys.VIRTUAL_FILE) ?: return
         val uriFile = Path(virtualFile.path).toUri()
 
         val request = PreviewMessageRequest(servers.first(), uriFile.toString(), componentName)
-        (servers.first() as LspServer).requestExecutor.sendRequestSync(request)
+        servers.first().requestExecutor.sendRequestSync(request)
 
         notification?.expire()
     }