+++ src/main/kotlin/dev/slint/ideaplugin/ide/widgets/SlintStatusBarWidgetFactory.kt
@@ -1,0 +1,20 @@
+package dev.slint.ideaplugin.ide.widgets
+
+import com.intellij.openapi.project.Project
+import com.intellij.openapi.wm.StatusBar
+import com.intellij.openapi.wm.StatusBarWidget
+import com.intellij.openapi.wm.StatusBarWidgetFactory
+
+class SlintStatusBarWidgetFactory : StatusBarWidgetFactory {
+    override fun getId(): String {
+        return "SlintWidget"
+    }
+
+    override fun getDisplayName(): String {
+        return "Slint"
+    }
+
+    override fun createWidget(project: Project): StatusBarWidget {
+        return SlintWidget(project)
+    }
+}